Update init.angler.rc - fix qmud
This change was applied a while ago to the 5X and never the 6P.for some reason.  The 6P has the same problem described googlesource link in my comments, so I think it's safe to assume that this small change got lost in the world of Google code and never was applied to the 6P as it should have been.

Trying to spread this small patch around to the xda devs so qmuxd can function properly and we can stop seeing it's logcat error spam.